new shop and delete all pictures completed

+++ SpringRestAPI/src/main/java/com/Java_M14/SpringRestAPI/service/PictureServiceImpl.java
@@ -5,6 +5,8 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.Java_M14.SpringRestAPI.dao.IPictureDAO;
@@ -14,10 +16,10 @@ import com.Java_M14.SpringRestAPI.dto.Shop;
 
 @Service
 public class PictureServiceImpl implements IPictureService {
-	
+
 	@Autowired
 	IPictureDAO iPictureDAO;
-	
+
 	@Autowired
 	IShopDAO iShopDAO;
 
@@ -26,8 +28,7 @@ public class PictureServiceImpl implements IPictureService {
 		Optional<Shop> shop = iShopDAO.findById(id);
 		return iPictureDAO.findByShop(shop);
 	}
-	
-	
+
 	@Override
 	public Picture addPicture(Long shopId, Map<String, String> pictureInfo) {
 		Picture newPicture = new Picture();
@@ -37,17 +38,21 @@ public class PictureServiceImpl implements IPictureService {
 		return iPictureDAO.save(newPicture);
 	}
 
-
 	@Override
-	public void deleteAllPictures(Long id) {
+	public ResponseEntity<String> deleteAllPictures(Long id) {
 		Optional<Shop> shop = iShopDAO.findById(id);
-		List <Picture> pictures = iPictureDAO.findByShop(shop);
-		for (Picture picture : pictures) {
-			iPictureDAO.delete(picture);
-		}
+		List<Picture> pictures = iPictureDAO.findByShop(shop);
 		
+		if (pictures.isEmpty()) {
+//			return new ResponseEntity<String>("EMpty", HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
+		} else {
+			for (Picture picture : pictures) {
+				iPictureDAO.delete(picture);
+			}
+			return new ResponseEntity<String>(HttpStatus.OK);
+		}
+
 	}
-
-
 
 }
